+++ alex_nb_web/src/main/java/nadrabank/dao/CreditHistoryDao.java
@@ -3,7 +3,6 @@ package nadrabank.dao;
 import nadrabank.domain.CreditHistory;
 import nadrabank.queryDomain.CreditAccPriceHistory;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 public interface CreditHistoryDao {
@@ -20,7 +19,7 @@ public interface CreditHistoryDao {
 
     List getLotIdHistoryByCredit(String inn, Long idBars);
 
-    BigDecimal getPriceByLotIdHistory(Long id, Long lotId);
+    Integer getPriceByLotIdHistory(Long id, Long lotId);
 
     List<CreditAccPriceHistory> getDateAndAccPriceHistoryByCredit(Long id);
 }
